searchByTags moved to picCtrl.js
