+++ server.js
@@ -110,32 +110,7 @@ App.post('/api/pic', PicController.create);
 App.get('/api/pic', function(req, res) {
 	res.status(200).json(imgs);
 });
-App.post('/api/tagPictures', function(req, res) {
-	var tempArr = req.body.tags
-	var searchArr = []
-	console.log(tempArr)
-	for(var i =0; i < tempArr.length; i++) {
-		Tags.find({title: tempArr[i]})
-			.select("_id")
-			.exec()
-			.then(function(id) {
-				id.forEach(function(item) {
-					searchArr.push(item._id)
-					console.log(searchArr)
-				})
-			Pics.find({ "tags": { "$all": searchArr}})
-			.select('_id imageUrl upvotes tags')
-			.populate('tags')
-			.sort('-upvotes')
-			.limit(50)
-			.exec()
-			.then(function(results) {
-				res.status(200).json(results)
-			})
-			})
-	}
-
-})
+App.post('/api/searchTagPictures', PicController.searchByTags);
 
 
 App.get("/api/profile", isAuthed, UserController.profile);
